Refactor: move bottom sheet
