+++ lib/widgets/tasks/task_form.dart
@@ -1,6 +1,5 @@
 // ignore_for_file: avoid_dynamic_calls
 
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
@@ -12,6 +11,7 @@ import 'package:lotti/classes/task.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/widgets/journal/editor/editor_widget.dart';
 import 'package:lotti/widgets/journal/entry_tools.dart';
+import 'package:lotti/widgets/misc/duration_bottom_sheet.dart';
 
 class TaskForm extends StatefulWidget {
   const TaskForm({
@@ -207,87 +207,6 @@ class TaskStatusLabel extends StatelessWidget {
         softWrap: false,
         style: inputStyle(),
       ),
-    );
-  }
-}
-
-class DurationBottomSheet extends StatefulWidget {
-  const DurationBottomSheet(this.initial, {super.key});
-
-  final Duration? initial;
-
-  @override
-  State<DurationBottomSheet> createState() => _DurationBottomSheetState();
-}
-
-class _DurationBottomSheetState extends State<DurationBottomSheet> {
-  Duration? duration;
-
-  @override
-  void initState() {
-    duration = widget.initial;
-    super.initState();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final localizations = AppLocalizations.of(context)!;
-
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        Container(
-          padding: const EdgeInsets.symmetric(
-            horizontal: 20,
-            vertical: 10,
-          ),
-          color: styleConfig().primaryColor.withOpacity(0.3),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              TextButton(
-                onPressed: () {
-                  Navigator.pop(context);
-                },
-                child: Text(
-                  localizations.cancelButton,
-                  style: buttonLabelStyle().copyWith(
-                    color: styleConfig().secondaryTextColor,
-                  ),
-                ),
-              ),
-              TextButton(
-                onPressed: () {
-                  Navigator.pop(context, duration);
-                },
-                child: Text(
-                  localizations.doneButton,
-                  style: buttonLabelStyle().copyWith(
-                    color: styleConfig().primaryColor,
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-        CupertinoTheme(
-          data: CupertinoThemeData(
-            textTheme: CupertinoTextThemeData(
-              pickerTextStyle: formLabelStyle().copyWith(
-                fontSize: fontSizeLarge,
-                fontWeight: FontWeight.w300,
-              ),
-            ),
-          ),
-          child: CupertinoTimerPicker(
-            onTimerDurationChanged: (Duration value) {
-              duration = value;
-            },
-            initialTimerDuration: widget.initial ?? Duration.zero,
-            mode: CupertinoTimerPickerMode.hm,
-          ),
-        ),
-      ],
     );
   }
 }
